// src/generated/models/New_personbolosModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const New_personbolosstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type New_personbolosstatecode = keyof typeof New_personbolosstatecode;
export const New_personbolosstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type New_personbolosstatuscode = keyof typeof New_personbolosstatuscode;

export interface New_personbolosBase {
  importsequencenumber?: number;
  new_age?: string;
  new_aka?: string;
  new_bolostatus?: string;
  new_bolotype?: string;
  new_casedetails?: string;
  new_casenumber?: string;
  new_city?: string;
  new_eyecolor?: string;
  new_firstname?: string;
  new_haircolor?: string;
  new_height?: string;
  new_lastname?: string;
  new_middlename?: string;
  new_name?: string;
  new_ownername?: string;
  new_personboloid: string;
  new_photourl?: string;
  new_race?: string;
  new_state?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: New_personbolosstatecode;
  statuscode?: New_personbolosstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface New_personbolos extends New_personbolosBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}