// src/generated/models/New_vehiclebolosModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const New_vehiclebolosstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type New_vehiclebolosstatecode = keyof typeof New_vehiclebolosstatecode;
export const New_vehiclebolosstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type New_vehiclebolosstatuscode = keyof typeof New_vehiclebolosstatuscode;

export interface New_vehiclebolosBase {
  importsequencenumber?: number;
  new_bolostatus?: string;
  new_bolotype?: string;
  new_casedetails?: string;
  new_casenumber?: string;
  new_city?: string;
  new_name?: string;
  new_ownername?: string;
  new_photourl?: string;
  new_platenumber?: string;
  new_platestate?: string;
  new_state?: string;
  new_vehicleboloid: string;
  new_vehiclecolor?: string;
  new_vehiclemake?: string;
  new_vehiclemodel?: string;
  new_vehicleyear?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: New_vehiclebolosstatecode;
  statuscode?: New_vehiclebolosstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface New_vehiclebolos extends New_vehiclebolosBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}